Refactor validation and smart constructors
Validation was previously done via a single smart constructor
`createInitialState` which took raw Int (number of tries) and String
(word to guess), and returned an Either containing either a list of
errors, or the initial RunningState.

This was a bit tedious to work with, in particular when trying to define
the user interface where we wished we could validate the number of tries
and the word independently; this would allow us, e.g., to filter the
list of words earlier instead of waiting for the initial state creation
to find out wether we have a good word or not.

Changes:
 - newType Tries to encapsulate a number of tries, and its associated
smart constructor
 - newType Word to encapsulate a good word, and its associated smart
constructor
 - newType Guess which represent a valid Guess, and its associated
smart constructor. By using this, we can enforce that a Guess is always
a lowerCase Char, which simplifies the logic a bit.
 - the lowerCaseWordToGuess has been removed, as it added complexity
without being that useful.